Created undoRedo files Linked list implementation is working correctly. Currently commit_move is not working properly (a dumb error with memory allocation of int ** boards), and undoRedo function for traversing the list is not yet implemented. Also added copy_board function to gameStruct.
FUTURE PLAN: after finishing with undoRedo functionality and testing, will test stackTracking a bit more extensively (specifically make
sure that the correct number of solutions is given when the board is partially/entirely full, or when it contains mistakes, visible or
otherwise). After that I will go over all of my files, and change local variables and argument names to conform to _underscore_ naming.
After that I'm out of work. Will probably start reading and messing around with Gurobi.