Fixed autocomplete bug where cells changed by autocomplete were affecting the number of possibilities for cells accessed later during the autocomplete, causing more cells to be autofilled than necessary. In order to buffer autocomplete values quickly, I reused a the stack implementation from stackTracking(*), and removed some superfluous comments from stackTracking.c while at it.
(*) in order to use the stack implementation in autocomplete module, access to the stack functions had to be given to all modules which
include stackTracking.h. This could potentially clash with the instructions in "CODING" section of the final project document, then again
this is reusing the stack code to great effect. Consider solutions.