Fixed an index bug in check_block_for_errors
The error was in row 121 of gameStruct.c, where the row index of the first reference cell (indexed with ri, rj)
was set to be `ri = ci + (cj + 1)/gptr->cols`. This was supposed to make ri equal ci unless cj was the last column
in the block. This did not take into consideration the fact that if this is not the first block in a row, then cj
is always greater than gptr->cols. In order to fix this problem, I made the check relative to sj, so it is now:

`ri = ci + (cj - sj + 1)/gptr->cols`

This seems to have solved the problem, and I added a more in depth explanation for the indexing and how it should be
like, but the indexing is still very complex.