Small review of undoRedo and userOps: I removed the check (*current)->next == NULL from append function. This check is apparently to prevent calling the function terminate on a null value, but since the test of the while loop in terminate is whether the current node is NULL or not, calling terminate on a null value will simply do nothing.
Questions:
1. Is the argument isFirst really necessary?
What's written in the comment about it:
"""
set previous of new node as current, and next as NULL.
this way, the previous pointer of the first node in an
undoRedo list points to itself, and the next pointer of
the last node points to NULL
"""
It seems to me this can be solved simply by calling append with a NULL value on current will solve the problem and make things easier.

2. I don't think calling is_fixed_erroneous necessary when going to edit mode. In edit mode, no value is fixed, so there is no need to call this function. Instead you can remove all fixed flags and check for errors regularly.

Those questions are not about any bug in particular, just my notes from the code. In any case, I'm moving ahead to guess function.