Fixed another bug in check_block_for_errors, where values in a block were compared to cells in former blocks in the same row. In order to fix the problem I changed the initial assignment of rj from:
rj = (ri == ci) ? (cj + 1) % gptr->cols : sj;

to:

rj = (ri == ci) ? (cj + 1) % (sj + gptr->cols) : sj;

In the previous version, if (ri==ci) in a block which is not the first it's row then (cj + 1) % gptr->cols would assign to rj a column value from the first block in the row. The fix makes sure all ri and rj assignemts stay in the same block.