First compilation build of the entire project
Contains changes to most files, including (but not limited to) moving all include and define lines into header files, and fixing includes to be according to the dependancy graph.

Project does not currently compile. Next task is to change userOp to be in a switch instead of a generic function.