Added automated test files:
Put both these files in the sudoku-console folder, and put the tester folders from IO.zip in a folder called tests (which will be inside the sudoku-console folder).
Then, get one of the boards from test_file_inputs (called inX), put it in the sudoku-console folder, and change it's name to in.
Last, change the `setenv` lines in ft.tcshrc to your gurobi `setenv` lines.

After running test.sh, a test_results folder will appear inside tests, and will contain a file for each test, which will have both input and output of that test.