Changed single cell value constraints
The constraints now apply to all cell values, valid or invalid, to make sure only one value is picked when generating a board.

Also, to make sure invalid values with 0.0 coefficient in the objective function are never picked, the model is always set to maximization, this should also reflect positively on guess.