Fixed validate empty cells deallocation bug
Fixed a bug where freeing the empty cell map was done with sideLength size instead of the number of empty cells. This caused gurobi_ilp function to crash when validating/hinting/saving a board with less than `siddeLength` empty cells.
Also changed main so that mark errors gets a default value of 1 (it should according to instructions).